+++ src/main/java/scanner/runner/logging/LoggerModule.java
@@ -4,13 +4,22 @@ import lombok.extern.slf4j.Slf4j;
 import scanner.runner.Target;
 import scanner.runner.logging.obj.LoggerInput;
 
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.util.Set;
 
+/**
+ * Модуль логирования результата анализа.
+ *
+ * @author inkarnadin
+ * on 08-01-2023
+ */
 @Slf4j
 public class LoggerModule {
 
+    /**
+     * Метод выполнения записи результата в лог.
+     *
+     * @param input входной объект-обертка для логирования результата
+     */
     public void execute(LoggerInput input) {
         Set<Target> targets = input.getTargets();
         targets.stream()