+++ src/main/java/scanner/runner/unlock/Credentials.java
@@ -1,4 +1,4 @@
-package scanner.runner.breaking;
+package scanner.runner.unlock;
 
 import lombok.Getter;