+++ src/main/java/scanner/runner/unlock/UnlockModule.java
@@ -1,27 +1,27 @@
-package scanner.runner.breaking;
+package scanner.runner.unlock;
 
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import scanner.runner.Target;
-import scanner.runner.breaking.brute.BruteScanner;
-import scanner.runner.breaking.cve.CVEScanner;
-import scanner.runner.breaking.obj.BreakInput;
-import scanner.runner.breaking.obj.BreakOutput;
-import scanner.runner.breaking.repeat.RepeatScanner;
+import scanner.runner.unlock.brute.BruteScanner;
+import scanner.runner.unlock.cve.CVEScanner;
+import scanner.runner.unlock.obj.UnlockInput;
+import scanner.runner.unlock.obj.UnlockOutput;
+import scanner.runner.unlock.repeat.RepeatScanner;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import static scanner.runner.logging.LoggingUtility.showSubtaskPercentCompletion;
+import static scanner.runner.logging.LoggingUtility.printSubtaskPercentCompletion;
 
 /**
- * Модуль разблокировки с помощью перебора.
+ * Модуль разблокировки доступа к устройству.
  *
  * @author inkarnadin
  * on 02-10-2022
  */
 @Slf4j
-public class BreakerModule {
+public class UnlockModule {
 
     private final CVEScanner cveScanner = new CVEScanner();
     private final BruteScanner bruteScanner = new BruteScanner();
@@ -33,7 +33,7 @@ public class BreakerModule {
      * @param input входной объект-обертка
      * @return выходной объект-обертка
      */
-    public BreakOutput execute(@NotNull BreakInput input) {
+    public UnlockOutput execute(@NotNull UnlockInput input) {
         Set<Target> result = new HashSet<>();
         try {
             Set<Target> targets = input.getAddresses();
@@ -47,14 +47,14 @@ public class BreakerModule {
                 Target resultItem = cveScanner.scanning(host);
                 if (resultItem.isBroken()) {
                     result.add(resultItem);
-                    showSubtaskPercentCompletion(++i, targets.size());
+                    printSubtaskPercentCompletion(++i, targets.size());
                     continue;
                 }
 
                 // brute
                 resultItem = bruteScanner.scanning(host, passwords.toArray(new String[0]));
                 result.add(resultItem);
-                showSubtaskPercentCompletion(++i, targets.size());
+                printSubtaskPercentCompletion(++i, targets.size());
             }
 
             //repeat
@@ -63,7 +63,7 @@ public class BreakerModule {
         } catch (Exception xep) {
             log.error("error during brute ip range [{}]: {}", input.getAddresses(), xep.getMessage());
         }
-        return BreakOutput.builder()
+        return UnlockOutput.builder()
                 .targets(result)
                 .build();
     }